Ajout des tests

+++ libasm.h
@@ -6,7 +6,7 @@
 /*   By: nguiard <nguiard@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/21 09:12:39 by nguiard           #+#    #+#             */
-/*   Updated: 2022/05/11 15:27:51 by nguiard          ###   ########.fr       */
+/*   Updated: 2023/04/24 13:09:03 by nguiard          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 # include <stddef.h>
 
-int			ft_strlen(const char *s);
+long int	ft_strlen(const char *s);
 char		*ft_strcpy(char *dest, const char *src);
 int			ft_strcmp(char *s1, char *s2);
 size_t		ft_write(int fd, char *addr, int count);
